# Optimize RedBlackTree-based immutable collections