… src/library/scala/collection/immutable/TreeSet.scala
… src/library/scala/math/Ordering.scala
… test/benchmarks/src/main/scala/scala/collection/immutable/RedBlackTreeBenchmark.scala
… test/scalacheck/redblacktree.scala