- Some improved inlining in RedBlackTree
- Remove unnecessary `contains` check before deleting from TreeSet or TreeMap
- Avoid recreating wrapper objects when appending to and deleting from TreeSets and TreeMaps without a change
- Bulk building / appending without creating intermediate TreeSets or TreeMaps
- Building TreeSets and TreeMaps directly without a Builder
- Use null instead of () as dummy value in TreeSet

Here's some performance data from Java 8 with HotSpot on x64. Getting reliable results was not always easy. I ran the benchmarks with `bench/jmh:run -jvmArgs "-Xms128M -Xmx128M" scala.collection.immutable.RedBlackTreeBenchmark`. Giving it lots of memory led to unreliable data, with optimizations kicking in very late only when there was GC pressure, so you'd need  a huge number of warmup iterations.

Building from a source in the correct order and in random order, in 2.13.x:
```
[info] Benchmark                          (size)  Mode  Cnt        Score       Error  Units
[info] RedBlackTreeBenchmark.build             0  avgt   60       15.032 ±     0.084  ns/op
[info] RedBlackTreeBenchmark.build             1  avgt   60       24.781 ±     0.111  ns/op
[info] RedBlackTreeBenchmark.build            10  avgt   60      741.863 ±     4.457  ns/op
[info] RedBlackTreeBenchmark.build           100  avgt   60    11803.651 ±    68.646  ns/op
[info] RedBlackTreeBenchmark.build          1000  avgt   60   178889.209 ±   947.915  ns/op
[info] RedBlackTreeBenchmark.build         10000  avgt   60  2641918.688 ± 13482.549  ns/op
[info] RedBlackTreeBenchmark.buildRandom       0  avgt   60       27.135 ±     0.155  ns/op
[info] RedBlackTreeBenchmark.buildRandom       1  avgt   60       44.238 ±     0.583  ns/op
[info] RedBlackTreeBenchmark.buildRandom      10  avgt   60      706.818 ±    15.026  ns/op
[info] RedBlackTreeBenchmark.buildRandom     100  avgt   60    11097.576 ±    87.879  ns/op
[info] RedBlackTreeBenchmark.buildRandom    1000  avgt   60   242717.013 ±  1577.754  ns/op
[info] RedBlackTreeBenchmark.buildRandom   10000  avgt   60  3932147.592 ± 23255.622  ns/op
```

With this change (avoiding the unnecessary intermediate `TreeSet` objects):
```
[info] RedBlackTreeBenchmark.build             0  avgt   60        6.529 ±     0.036  ns/op
[info] RedBlackTreeBenchmark.build             1  avgt   60       11.644 ±     0.126  ns/op
[info] RedBlackTreeBenchmark.build            10  avgt   60      398.927 ±     2.394  ns/op
[info] RedBlackTreeBenchmark.build           100  avgt   60     8533.978 ±    50.420  ns/op
[info] RedBlackTreeBenchmark.build          1000  avgt   60   164994.754 ±  1869.387  ns/op
[info] RedBlackTreeBenchmark.build         10000  avgt   60  2253731.587 ± 11218.765  ns/op
[info] RedBlackTreeBenchmark.buildRandom       0  avgt   60       25.479 ±     0.160  ns/op
[info] RedBlackTreeBenchmark.buildRandom       1  avgt   60       32.923 ±     0.642  ns/op
[info] RedBlackTreeBenchmark.buildRandom      10  avgt   60      346.081 ±     2.612  ns/op
[info] RedBlackTreeBenchmark.buildRandom     100  avgt   60     8221.890 ±    43.662  ns/op
[info] RedBlackTreeBenchmark.buildRandom    1000  avgt   60   209767.283 ±  1639.279  ns/op
[info] RedBlackTreeBenchmark.buildRandom   10000  avgt   60  3668358.135 ± 51228.253  ns/op
```

Deleting all elements in random order in 2.13.x:
```
[info] RedBlackTreeBenchmark.drain             0  avgt   60        4.347 ±     0.032  ns/op
[info] RedBlackTreeBenchmark.drain             1  avgt   60       13.908 ±     0.078  ns/op
[info] RedBlackTreeBenchmark.drain            10  avgt   60      563.643 ±     3.989  ns/op
[info] RedBlackTreeBenchmark.drain           100  avgt   60    10659.737 ±    69.129  ns/op
[info] RedBlackTreeBenchmark.drain          1000  avgt   60   284227.686 ±  2179.154  ns/op
[info] RedBlackTreeBenchmark.drain         10000  avgt   60  4932893.507 ± 34018.351  ns/op
```

With this change (benefiting mostly from not performing the unnecessary `contains` checks):
```
[info] RedBlackTreeBenchmark.drain             0  avgt   60        4.334 ±     0.030  ns/op
[info] RedBlackTreeBenchmark.drain             1  avgt   60       13.726 ±     0.082  ns/op
[info] RedBlackTreeBenchmark.drain            10  avgt   60      486.583 ±     2.400  ns/op
[info] RedBlackTreeBenchmark.drain           100  avgt   60     9553.461 ±    54.927  ns/op
[info] RedBlackTreeBenchmark.drain          1000  avgt   60   220971.471 ±  1502.734  ns/op
[info] RedBlackTreeBenchmark.drain         10000  avgt   60  4115382.806 ± 27286.882  ns/op
```

Performance of iteration via `iterator` and `foreach` is unchanged.

I spent the better part of a week on this and the resulting set of changes is rather small. There's a lot more that I tried that didn't work:

- Removing `null` trees and using a proper `empty` tree object doesn't give you any real advantage. The way the red-black tree algorithms work you have to treat the empty tree specially in almost all cases so you may as well check for null. And even if you add an `empty` tree you'll still want to check for that by comparing it with `eq empty` because this is much faster than an `isEmpty` method on the tree itself.

- Removing polymorphic dispatch of `Tree` methods by making `Tree` final is promising. You have to encode the color somehow if it's not encoded in the class anymore. Adding an extra field is not a good option. The basic `Tree` type has a size of 32 bytes with no wasted space, just like the original `RedTree` and `BlackTree` types. Adding an extra field would increase the memory footprint by 25% (assuming 64-bit alignment), so I used the (otherwise unused) sign bit of the `count` field for the color:
  ```scala
    final class Tree[A, +B](
      @(`inline` @getter) final val key: A,
      @(`inline` @getter) final val value: B,
      @(`inline` @getter) final val left: Tree[A, B],
      @(`inline` @getter) final val right: Tree[A, B],
      @(`inline` @getter) final val _count: Int) // negative for black trees, positive for red trees
  ```
  This works rather well in *some* cases:
  ```
  [info] RedBlackTreeBenchmark.buildRandom       0  avgt   16       24.148 ±     0.250  ns/op
  [info] RedBlackTreeBenchmark.buildRandom       1  avgt   16       37.722 ±     0.615  ns/op
  [info] RedBlackTreeBenchmark.buildRandom      10  avgt   16      378.835 ±     7.607  ns/op
  [info] RedBlackTreeBenchmark.buildRandom     100  avgt   16     7688.842 ±    93.652  ns/op
  [info] RedBlackTreeBenchmark.buildRandom    1000  avgt   16   186845.275 ±  3112.246  ns/op
  [info] RedBlackTreeBenchmark.buildRandom   10000  avgt   16  3210497.905 ± 50008.618  ns/op
  ```
  Other benchmarks, in particular for iteration, showed decreased performance. Looking back now at the results I got when I did these tests, it's possible that they were skewed by the late optimization without limiting the heap size. I'll do another proper test run tonight. Maybe this encoding is better after all.

- Inlining some of the small methods in `TreeIterator`. This looks like an obvious win but it's not. Letting HotSpot take care of inlining is actually better in this case.

- Iteration instead of tail recursion in `foreach`. We can't enforce tail recursion (marking the method `@tailrec` is not an option because it is also used in non-tail position) but scalac still generates better code for the tail recursion than for an explicit loop.

- Using a Java implementation of the tree classes. The Scala version with the ``@(`inline` @getter)`` annotations is clunky, but so is Java. Performance is exactly the same.