@szeiger 2018-10-01
Here are the results from the alternate encoding. The performance regression in `iterator` that I saw earlier was indeed caused by bad measurement.  Iteration with `iterator` and `foreach` is basically the same as in 2.13.x and in this PR. `drain` is a bit slower than the version in this PR:
```
[info] RedBlackTreeBenchmark.drain             0  avgt   60        4.375 ±     0.035  ns/op
[info] RedBlackTreeBenchmark.drain             1  avgt   60       13.876 ±     0.088  ns/op
[info] RedBlackTreeBenchmark.drain            10  avgt   60      503.655 ±     2.433  ns/op
[info] RedBlackTreeBenchmark.drain           100  avgt   60    10057.843 ±    57.836  ns/op
[info] RedBlackTreeBenchmark.drain          1000  avgt   60   239275.018 ±  1582.677  ns/op
[info] RedBlackTreeBenchmark.drain         10000  avgt   60  4291043.901 ± 25764.321  ns/op
```
Building in random order is faster than this PR, building in correct order is slower:
```
[info] RedBlackTreeBenchmark.build             0  avgt   60        6.511 ±     0.031  ns/op
[info] RedBlackTreeBenchmark.build             1  avgt   60       19.576 ±     0.090  ns/op
[info] RedBlackTreeBenchmark.build            10  avgt   60      431.087 ±     3.340  ns/op
[info] RedBlackTreeBenchmark.build           100  avgt   60     9434.790 ±    56.822  ns/op
[info] RedBlackTreeBenchmark.build          1000  avgt   60   186017.401 ±  1021.874  ns/op
[info] RedBlackTreeBenchmark.build         10000  avgt   60  2646389.650 ± 21761.592  ns/op
[info] RedBlackTreeBenchmark.buildRandom       0  avgt   60       25.820 ±     0.109  ns/op
[info] RedBlackTreeBenchmark.buildRandom       1  avgt   60       40.536 ±     0.269  ns/op
[info] RedBlackTreeBenchmark.buildRandom      10  avgt   60      406.591 ±     2.317  ns/op
[info] RedBlackTreeBenchmark.buildRandom     100  avgt   60     8314.624 ±    45.755  ns/op
[info] RedBlackTreeBenchmark.buildRandom    1000  avgt   60   195968.450 ±  1292.088  ns/op
[info] RedBlackTreeBenchmark.buildRandom   10000  avgt   60  3324854.275 ± 22224.404  ns/op
```
All in all it doesn't look like a worthwhile improvement.