@szeiger 2018-10-02
Updated with better Builders (which avoid wrapping the `RB.Tree` between additions) and efficient bulk operations.

Here are the benchmark results for the bulk `union` operation. I didn't benchmark `intersect` and `diff` but they are also implemented and tested.

Base line: Old version of this PR minus the `concat` override. This should be the same as or slightly faster than the current 2.13.x:
```
[info] Benchmark                    (size)  Mode  Cnt         Score       Error  Units
[info] RedBlackTreeBenchmark.union       0  avgt   20       137.777 ±     1.718  ns/op
[info] RedBlackTreeBenchmark.union       1  avgt   20       154.537 ±     3.338  ns/op
[info] RedBlackTreeBenchmark.union      10  avgt   20      1808.558 ±    15.092  ns/op
[info] RedBlackTreeBenchmark.union     100  avgt   20     39168.882 ±   205.147  ns/op
[info] RedBlackTreeBenchmark.union    1000  avgt   20    838237.278 ±  5720.816  ns/op
[info] RedBlackTreeBenchmark.union   10000  avgt   20  10802699.255 ± 90692.569  ns/op
```

Old version of this PR including the `concat` override to avoid building intermediate wrappers:
```
[info] Benchmark                    (size)  Mode  Cnt        Score       Error  Units
[info] RedBlackTreeBenchmark.union       0  avgt   20        9.278 ±     0.058  ns/op
[info] RedBlackTreeBenchmark.union       1  avgt   20       76.321 ±     0.498  ns/op
[info] RedBlackTreeBenchmark.union      10  avgt   20     1032.589 ±     7.778  ns/op
[info] RedBlackTreeBenchmark.union     100  avgt   20    20490.622 ±    74.912  ns/op
[info] RedBlackTreeBenchmark.union    1000  avgt   20   443673.045 ±  1663.212  ns/op
[info] RedBlackTreeBenchmark.union   10000  avgt   20  5678857.481 ± 30469.581  ns/op
```

Optimized bulk `union`:
```
[info] Benchmark                    (size)  Mode  Cnt       Score      Error  Units
[info] RedBlackTreeBenchmark.union       0  avgt   20       6.615 ±    0.058  ns/op
[info] RedBlackTreeBenchmark.union       1  avgt   20      40.921 ±    1.035  ns/op
[info] RedBlackTreeBenchmark.union      10  avgt   20    1274.996 ±    6.919  ns/op
[info] RedBlackTreeBenchmark.union     100  avgt   20   10662.666 ±   73.296  ns/op
[info] RedBlackTreeBenchmark.union    1000  avgt   20   59756.993 ±  685.840  ns/op
[info] RedBlackTreeBenchmark.union   10000  avgt   20  313465.974 ± 3380.362  ns/op
```

Note that this is single-threaded but the algorithms are highly scalable. This might be a useful addition to parallel-collections.