@retronym 2018-10-02
@szeiger Great results!

It's useful also show the improvements in the `gc.alloc.rate.norm` counter by adding `-prof gc` to the JMH runs.