@retronym 2018-10-02
I think it would be worthwhile for us to implement `.equals` on our standard `Ordering` instances and change the optimization herein to be based on `==` rather than `eq`.  This would be needed for `Option` / `Iterable` / `TupleN`.