@szeiger 2018-10-04
I was just looking at the possibility of using mutable fields during construction. The caveat is that we'd have to switch to the alternate encoding with a single `Tree` class, otherwise there is no way to change the color of a node, the most essential change without which changing any of the other fields is not very useful.