@szeiger 2018-10-04
I just pushed another commit with further algorithmic improvements: We can build red-black trees very efficiently in linear time and fully immutable from ordered sets/maps (including `Range`, which is effectively an ordered set).

Here are the new results for building from a `Range`:
```
[info] Benchmark                          (size)  Mode  Cnt        Score       Error  Units
[info] RedBlackTreeBenchmark.build             0  avgt   20       44.315 ±     0.402  ns/op
[info] RedBlackTreeBenchmark.build             1  avgt   20       52.186 ±     1.208  ns/op
[info] RedBlackTreeBenchmark.build            10  avgt   20      139.171 ±     1.224  ns/op
[info] RedBlackTreeBenchmark.build           100  avgt   20     1040.310 ±     8.710  ns/op
[info] RedBlackTreeBenchmark.build          1000  avgt   20    10074.338 ±   164.073  ns/op
[info] RedBlackTreeBenchmark.build         10000  avgt   20   114782.362 ±  3064.179  ns/op
```
For the largest size this is a 20x improvement over building by incremental insertion.