@szeiger 2018-10-08
While looking at another PR that optimizes `filter` for `HashMap` I realized that we don't have a red-black tree specific version, so I implemented it based on `join`. As usual, the performance difference is huge:

Before:
```
[info] Benchmark                            (size)  Mode  Cnt        Score       Error  Units
[info] RedBlackTreeBenchmark.filterDrop          0  avgt   20       27.744 ±     0.277  ns/op
[info] RedBlackTreeBenchmark.filterDrop          1  avgt   20       43.919 ±     0.301  ns/op
[info] RedBlackTreeBenchmark.filterDrop         10  avgt   20       79.747 ±     1.354  ns/op
[info] RedBlackTreeBenchmark.filterDrop        100  avgt   20      374.714 ±     1.503  ns/op
[info] RedBlackTreeBenchmark.filterDrop       1000  avgt   20     3110.857 ±    23.053  ns/op
[info] RedBlackTreeBenchmark.filterDrop      10000  avgt   20    37306.189 ±  1281.405  ns/op
[info] RedBlackTreeBenchmark.filterKeep          0  avgt   20       27.636 ±     0.352  ns/op
[info] RedBlackTreeBenchmark.filterKeep          1  avgt   20       66.643 ±     0.868  ns/op
[info] RedBlackTreeBenchmark.filterKeep         10  avgt   20      439.471 ±    19.799  ns/op
[info] RedBlackTreeBenchmark.filterKeep        100  avgt   20     8527.503 ±    41.508  ns/op
[info] RedBlackTreeBenchmark.filterKeep       1000  avgt   20   172379.713 ±  1408.782  ns/op
[info] RedBlackTreeBenchmark.filterKeep      10000  avgt   20  2313769.548 ± 10185.289  ns/op
[info] RedBlackTreeBenchmark.filterPartial       0  avgt   20       27.436 ±     0.244  ns/op
[info] RedBlackTreeBenchmark.filterPartial       1  avgt   20       43.689 ±     0.413  ns/op
[info] RedBlackTreeBenchmark.filterPartial      10  avgt   20      197.619 ±     2.691  ns/op
[info] RedBlackTreeBenchmark.filterPartial     100  avgt   20     3804.073 ±    26.264  ns/op
[info] RedBlackTreeBenchmark.filterPartial    1000  avgt   20    78183.333 ±   408.709  ns/op
[info] RedBlackTreeBenchmark.filterPartial   10000  avgt   20  1151320.092 ±  7191.067  ns/op
```

After:
```
[info] Benchmark                            (size)  Mode  Cnt       Score      Error  Units
[info] RedBlackTreeBenchmark.filterDrop          0  avgt   20       3.397 ±    0.031  ns/op
[info] RedBlackTreeBenchmark.filterDrop          1  avgt   20       7.898 ±    0.138  ns/op
[info] RedBlackTreeBenchmark.filterDrop         10  avgt   20      36.673 ±    0.325  ns/op
[info] RedBlackTreeBenchmark.filterDrop        100  avgt   20     306.876 ±    2.876  ns/op
[info] RedBlackTreeBenchmark.filterDrop       1000  avgt   20    4180.397 ±   69.837  ns/op
[info] RedBlackTreeBenchmark.filterDrop      10000  avgt   20   35805.440 ±  748.316  ns/op
[info] RedBlackTreeBenchmark.filterKeep          0  avgt   20       3.702 ±    0.321  ns/op
[info] RedBlackTreeBenchmark.filterKeep          1  avgt   20       4.832 ±    0.095  ns/op
[info] RedBlackTreeBenchmark.filterKeep         10  avgt   20      49.131 ±    0.695  ns/op
[info] RedBlackTreeBenchmark.filterKeep        100  avgt   20     385.484 ±    4.185  ns/op
[info] RedBlackTreeBenchmark.filterKeep       1000  avgt   20    5399.643 ±   55.518  ns/op
[info] RedBlackTreeBenchmark.filterKeep      10000  avgt   20   51359.206 ±  609.722  ns/op
[info] RedBlackTreeBenchmark.filterPartial       0  avgt   20       3.505 ±    0.037  ns/op
[info] RedBlackTreeBenchmark.filterPartial       1  avgt   20       8.367 ±    0.183  ns/op
[info] RedBlackTreeBenchmark.filterPartial      10  avgt   20     182.781 ±    1.791  ns/op
[info] RedBlackTreeBenchmark.filterPartial     100  avgt   20    1748.557 ±   21.808  ns/op
[info] RedBlackTreeBenchmark.filterPartial    1000  avgt   20   26803.034 ±  293.217  ns/op
[info] RedBlackTreeBenchmark.filterPartial   10000  avgt   20  255889.476 ± 3419.415  ns/op
```

We didn't have an optimized `transform` method, either, even though it is trivial to implement and offers huge performance benefits, so I added one.

Before:
```
[info] Benchmark                            (size)  Mode  Cnt        Score       Error  Units
[info] RedBlackTreeBenchmark.transformAll        0  avgt   20       19.394 ±     0.368  ns/op
[info] RedBlackTreeBenchmark.transformAll        1  avgt   20       56.436 ±     0.661  ns/op
[info] RedBlackTreeBenchmark.transformAll       10  avgt   20      466.717 ±     6.555  ns/op
[info] RedBlackTreeBenchmark.transformAll      100  avgt   20     9562.954 ±   100.270  ns/op
[info] RedBlackTreeBenchmark.transformAll     1000  avgt   20   181299.465 ±  1556.040  ns/op
[info] RedBlackTreeBenchmark.transformAll    10000  avgt   20  2543622.330 ± 26071.983  ns/op
```

After:
```
[info] RedBlackTreeBenchmark.transformAll        0  avgt   20       3.476 ±    0.083  ns/op
[info] RedBlackTreeBenchmark.transformAll        1  avgt   20      16.578 ±    0.103  ns/op
[info] RedBlackTreeBenchmark.transformAll       10  avgt   20     115.138 ±    1.307  ns/op
[info] RedBlackTreeBenchmark.transformAll      100  avgt   20    1215.023 ±    7.633  ns/op
[info] RedBlackTreeBenchmark.transformAll     1000  avgt   20   14840.987 ±  214.410  ns/op
[info] RedBlackTreeBenchmark.transformAll    10000  avgt   20  149494.743 ± 1293.007  ns/op
```